Got 3idemo to work
Managed to get 3idemo to work, all that was needing to change was line 20, though for some reason, the raspberry pi did not register "database_data" but when i changed it to "DATABASE_data" in both 3idemo.py and database.py it ran smoothly.
i changed line 42 to make it only insert 1 set of data per run of the program.