Bluetooth database creator and device scanner
Working progress. The database file is a simple create a database to hold mac addresses of registered devices.
The blue scanner file is a work in progress program that will currently request username, devicename and will scan bluetooth and enter the details into the database and display a greeting the temp into the pi.